Made new GameLogic module
